refactor(ui): split V2 control-plane views
