+++ web/src/controlPlaneQueries.ts
@@ -1,0 +1,8 @@
+import type { QueryClient } from "@tanstack/react-query";
+
+export function invalidateControlPlane(queryClient: QueryClient) {
+  return Promise.all([
+    queryClient.invalidateQueries({ queryKey: ["tasks"] }),
+    queryClient.invalidateQueries({ queryKey: ["workers"] }),
+  ]);
+}
